refactor: UI Model의 패키지 경로 변경
