+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/domain/repository/TimelineRepository.kt
@@ -1,6 +1,6 @@
 package com.woowacourse.staccato.domain.repository
 
-import com.woowacourse.staccato.presentation.timeline.TimelineTravelUiModel
+import com.woowacourse.staccato.presentation.timeline.model.TimelineTravelUiModel
 
 interface TimelineRepository {
     fun loadTravels(): List<TimelineTravelUiModel>
