+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/timeline/TimelineViewModel.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.woowacourse.staccato.domain.repository.TimelineRepository
+import com.woowacourse.staccato.presentation.timeline.model.TimelineTravelUiModel
 
 class TimelineViewModel(private val repository: TimelineRepository) : ViewModel() {
     private val _travels = MutableLiveData<List<TimelineTravelUiModel>>()
