+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/timeline/adapter/LastTravelViewHolder.kt
@@ -2,7 +2,7 @@ package com.woowacourse.staccato.presentation.timeline.adapter
 
 import com.woowacourse.staccato.databinding.LayoutItemFragmentTimelineLastBinding
 import com.woowacourse.staccato.presentation.timeline.TimelineHandler
-import com.woowacourse.staccato.presentation.timeline.TimelineTravelUiModel
+import com.woowacourse.staccato.presentation.timeline.model.TimelineTravelUiModel
 
 class LastTravelViewHolder(
     private val binding: LayoutItemFragmentTimelineLastBinding,
