+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/timeline/adapter/MiddleTravelViewHolder.kt
@@ -2,7 +2,7 @@ package com.woowacourse.staccato.presentation.timeline.adapter
 
 import com.woowacourse.staccato.databinding.LayoutItemFragmentTimelineMiddleBinding
 import com.woowacourse.staccato.presentation.timeline.TimelineHandler
-import com.woowacourse.staccato.presentation.timeline.TimelineTravelUiModel
+import com.woowacourse.staccato.presentation.timeline.model.TimelineTravelUiModel
 
 class MiddleTravelViewHolder(
     private val binding: LayoutItemFragmentTimelineMiddleBinding,
