+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/timeline/model/TimelineTravelUiModel.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.staccato.presentation.timeline
+package com.woowacourse.staccato.presentation.timeline.model
 
 data class TimelineTravelUiModel(
     val travelId: Long,
