+++ android/Staccato_AN/app/src/main/res/layout/layout_item_fragment_timeline_middle.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="travel"
-            type="com.woowacourse.staccato.presentation.timeline.TimelineTravelUiModel" />
+            type="com.woowacourse.staccato.presentation.timeline.model.TimelineTravelUiModel" />
 
         <variable
             name="eventHandler"
